refactor(inventory): Phase 2a move integration event payloads to Contracts/
- 8 payload records moved to Contracts/Integration/ under namespace
  Inventory.Service.Contracts.Integration.
- EventHandlers/ stay in IntegrationEvents/ for now; Phase 6 dissolves them.
