+++ inventory-microservice/Inventory.Service/Contracts/Integration/LowStockEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Inventory.Service.IntegrationEvents;
+namespace Inventory.Service.Contracts.Integration;
 
 public record LowStockEvent(
     int ProductId,
