+++ inventory-microservice/Inventory.Service/Contracts/Integration/ProductCreatedEvent.cs
@@ -1,5 +1,5 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Inventory.Service.IntegrationEvents;
+namespace Inventory.Service.Contracts.Integration;
 
 public record ProductCreatedEvent(int ProductId, string Name, decimal Price) : Event;
