+++ inventory-microservice/Inventory.Service/Contracts/Integration/StockAdjustedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Inventory.Service.IntegrationEvents;
+namespace Inventory.Service.Contracts.Integration;
 
 public record StockAdjustedEvent(
     int ProductId,
