+++ inventory-microservice/Inventory.Service/Contracts/Integration/StockCommittedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Inventory.Service.IntegrationEvents;
+namespace Inventory.Service.Contracts.Integration;
 
 public record CommittedItem(int ProductId, int WarehouseId, int Quantity);
