+++ inventory-microservice/Inventory.Service/Contracts/Integration/StockReservedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Inventory.Service.IntegrationEvents;
+namespace Inventory.Service.Contracts.Integration;
 
 public record ReservedItem(int ProductId, int WarehouseId, int Quantity);
