+++ inventory-microservice/Inventory.Service/Endpoints/InventoryApiEndpoints.cs
@@ -2,8 +2,8 @@ using ECommerce.Shared.Infrastructure.EventBus;
 using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.Observability.Metrics;
 using Inventory.Service.ApiModels;
+using Inventory.Service.Contracts.Integration;
 using Inventory.Service.Infrastructure.Data;
-using Inventory.Service.IntegrationEvents;
 using Inventory.Service.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
