+++ inventory-microservice/Inventory.Tests/Api/ReserveStockCommandHandlerTests.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
 using ECommerce.Shared.IntegrationEvents.Commands;
-using Inventory.Service.IntegrationEvents;
+using Inventory.Service.Contracts.Integration;
 using Inventory.Service.IntegrationEvents.EventHandlers;
 using Inventory.Service.Models;
 using Microsoft.Extensions.DependencyInjection;
